backend/requirements.txt
httpcore
fastapi[all]
pyyaml